+++ axelor-base/src/main/java/com/axelor/apps/base/web/CurrencyConversionLineController.java
@@ -19,6 +19,7 @@ package com.axelor.apps.base.web;
 
 import java.lang.invoke.MethodHandles;
 import java.math.BigDecimal;
+import java.net.MalformedURLException;
 import java.util.Map;
 
 import org.slf4j.Logger;
@@ -31,11 +32,14 @@ import com.axelor.apps.base.db.repo.CurrencyRepository;
 import com.axelor.apps.base.exceptions.IExceptionMessage;
 import com.axelor.apps.base.service.CurrencyConversionService;
 import com.axelor.apps.base.service.app.AppBaseService;
+import com.axelor.exception.AxelorException;
 import com.axelor.i18n.I18n;
 import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
 import com.axelor.rpc.Context;
 import com.google.inject.Inject;
+
+import wslite.json.JSONException;
 
 public class CurrencyConversionLineController {
 	
@@ -76,7 +80,7 @@ public class CurrencyConversionLineController {
 	}
 	
 	@SuppressWarnings("unchecked")
-	public void convert(ActionRequest request, ActionResponse response) {
+	public void convert(ActionRequest request, ActionResponse response) throws MalformedURLException, JSONException, AxelorException {
 		Context context = request.getContext();
 		Currency fromCurrency = null;
 		Currency toCurrency = null;